+++ CHANGELOG.rst
@@ -34,6 +34,7 @@ Bugfixes
 * Fix terminal colors on Windows.
 * Fix parsing Markdown metadata from solution.
 * Fix parsing Markdown solution data with '---' string.
+* Fix parsing Markdown solution brief attribute.
 * All commits in `0.11.0`_.
 
 Security